requirements.txt
django
django-tailwind[reload]
djangorestframework
django-utils